Backend/requirements.txt
fastapi>=0.103.0
uvicorn>=0.23.2
sqlalchemy>=2.0.20
pydantic>=2.3.0
psycopg2-binary>=2.9.7
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
yfinance>=0.2.28
nsepy>=0.8
pandas>=2.1.0
python-dotenv>=1.0.0
redis>=4.6.0
python-multipart>=0.0.6
pytest>=7.4.0
httpx>=0.24.1
supabase>=0.7.1
pytz>=2023.3
